Common Function written
