+++ apps/front-end/src/component/BaseInput.js
@@ -45,6 +45,19 @@ export function AddButton({ icon, iconType, ...btnProps }) {
   );
 }
 
+export const DocumentStatus = (docStatus) => {
+  if (!docStatus) {
+    return false;
+  } else {
+    const JsonConvert = docStatus ? JSON.parse(docStatus) : {};
+    // Check if all values are either "complete" or "not_applicable"
+    const isAllCompleteOrNotApplicable = Object.values(JsonConvert).every(
+      (value) => value === "complete" || value === "not_applicable"
+    );
+    return isAllCompleteOrNotApplicable ? true : false;
+  }
+};
+
 export function RemoveButton({ icon, iconType, ...btnProps }) {
   const { t } = useTranslation();
   return (
@@ -629,6 +642,7 @@ const validator = customizeValidator({
     MobileNumber: /^[6-9]\d{8}9$/,
   },
 });
+
 
 const widgets = {
   RadioBtn,
